#152 Refactor: split code into 2 class PreventEvents and InputBreakDown
