+++ javascript/big-practice/src/scripts/helper/preventEvents.js
@@ -1,0 +1,10 @@
+export default class PreventEvents {
+  /**
+   * @description function stop overlay bubbling event of 2 input note form
+   */
+  static stopEvents(element) {
+    element.addEventListener('click', (e) => {
+      e.stopPropagation();
+    });
+  }
+}
